First commit
First version. Currently scrapes all of the property data for Vancouver houses for sale on the Remax.ca website. 

Code is still very basic. Next step is separating the beds/baths column in the written excel file and finding the neighbourhoods each home is located in so that we can analyse the distribution of house prices between different neighbourhoods based on number of bedrooms and baths. This could be done easily in excel but I will be writing it all in Python for a challenge. We can then do the analysis and visualisation. 

The number of pages is also hard-coded in. This could be improved.

I would also like to find data on square footage of each property somehow.